[refactor] * refactor conftests * change yaml file * add allure marks
