+++ pytest.ini
@@ -7,9 +7,4 @@ filterwarnings =
     ignore: HTTPResponse.getheader()
     ignore: '_request_ctx_stack' is deprecated
 markers =
-    login_guest: tests for registered users
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    serial
-    test_user: test for users
     unit: unit-tests only
-    smoke: basic functionality
