+++ tests/conftest.py
@@ -9,10 +9,11 @@ from allure_commons.types import AttachmentType
 from faker import Faker
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service as ChromeService
+from selenium.webdriver.remote.webdriver import WebDriver
 from webdriver_manager.chrome import ChromeDriverManager
 
 from app import db, create_app
-from app.models import User, Service, Review
+from app.models import User, Service
 from config import TestConfig
 from tests.helpers.helpers import generate_phone_number
 from tests.ui_tests.pages.main_page import MainPage
@@ -58,14 +59,14 @@ def driver(request) -> Generator:
 
 
 @pytest.fixture
-def main_page(driver):
+def main_page(driver) -> MainPage:
     """Открывает главную страницу"""
     page = MainPage(driver)
     page.open()
     return page
 
 
-def screenshot(browser, name: str) -> None:
+def screenshot(browser: WebDriver, name: str) -> None:
     """
     Gets a screenshot and attaches it to the report
     """
@@ -83,32 +84,8 @@ def user() -> User:
 
 
 @pytest.fixture
-def service(user):
-    """Creates a user instance for testing"""
-    db.session.add(user)
-    db.session.commit()
-    service = Service(service1='A jar of honey', service_date=datetime.now(),
-                      service_time='10:00', user_id=user.id)
-    db.session.add(service)
-    db.session.commit()
-
-
-@pytest.fixture
-def new_service(user):
+def new_service() -> Service:
     """Returns a service for testing"""
     service = Service(service1='Chicken Burger', service2='Pepsi', service3='Delivery',
-                      service_date=datetime.now(), service_time='14:00', user_id=user.id)
+                      service_date=datetime.now(), service_time='14:00', user_id='1')
     return service
-
-
-@pytest.fixture
-def review(user):
-    """Creates a service for testing"""
-
-    db.session.add(user)
-    db.session.commit()
-    rev1 = Review(author=user.username, text='Sweeter than honey!', rating='Awesome!', author_id=user.id)
-    rev2 = Review(author=user.username, text='Bitter than Kopalkhen', rating='Terrible!', author_id=user.id)
-    rev3 = Review(author=user.username, text='Not bad', rating='So-so', author_id=user.id)
-    db.session.add_all([rev1, rev2, rev3])
-    db.session.commit()
